create common component RoundIconComponent
